# tools/st/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"io"
)

// Ref: #/components/schemas/Error
type Error struct {
	Message string `json:"message"`
}

// GetMessage returns the value of Message.
func (s *Error) GetMessage() string {
	return s.Message
}

// SetMessage sets the value of Message.
func (s *Error) SetMessage(val string) {
	s.Message = val
}

type GetFileInternalServerError Error

func (*GetFileInternalServerError) getFileRes() {}

type GetFileNotFound Error

func (*GetFileNotFound) getFileRes() {}

type GetFileOK struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s GetFileOK) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

func (*GetFileOK) getFileRes() {}

type HealthOK struct {
	Status OptString `json:"status"`
}

// GetStatus returns the value of Status.
func (s *HealthOK) GetStatus() OptString {
	return s.Status
}

// SetStatus sets the value of Status.
func (s *HealthOK) SetStatus(val OptString) {
	s.Status = val
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

type UploadFileBadRequest Error

func (*UploadFileBadRequest) uploadFileRes() {}

type UploadFileInternalServerError Error

func (*UploadFileInternalServerError) uploadFileRes() {}

type UploadFileReq struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s UploadFileReq) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

// Ref: #/components/schemas/UploadResponse
type UploadResponse struct {
	Filename string `json:"filename"`
	Size     int    `json:"size"`
	Message  string `json:"message"`
}

// GetFilename returns the value of Filename.
func (s *UploadResponse) GetFilename() string {
	return s.Filename
}

// GetSize returns the value of Size.
func (s *UploadResponse) GetSize() int {
	return s.Size
}

// GetMessage returns the value of Message.
func (s *UploadResponse) GetMessage() string {
	return s.Message
}

// SetFilename sets the value of Filename.
func (s *UploadResponse) SetFilename(val string) {
	s.Filename = val
}

// SetSize sets the value of Size.
func (s *UploadResponse) SetSize(val int) {
	s.Size = val
}

// SetMessage sets the value of Message.
func (s *UploadResponse) SetMessage(val string) {
	s.Message = val
}

func (*UploadResponse) uploadFileRes() {}