# tools/st/openapi.yaml
openapi: 3.0.0
info:
  title: File Server API
  version: 1.0.0
  description: Simple file serving API
paths:
  /posts/{filename}:
    get:
      operationId: getFile
      summary: Get a file by filename
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Name of the file to retrieve
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: uploadFile
      summary: Upload a file
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Name of the file to upload
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /health:
    get:
      operationId: health
      summary: Health check endpoint
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    UploadResponse:
      type: object
      properties:
        filename:
          type: string
        size:
          type: integer
        message:
          type: string
      required:
        - filename
        - size
        - message